% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distSquared}
\alias{distSquared}
\title{Radial basis function kernel matrix}
\usage{
distSquared(A, B)
}
\arguments{
\item{X}{an n x d matrix or numeric vector}

\item{Y}{an n x d matrix or numeric vector}
}
\value{
an n x n distance matrix
}
\description{
Radial basis function kernel matrix
}

