\name{KDRcpp-package}
\alias{KDRcpp-package}
\alias{KDRcpp}
\docType{package}
\title{
\packageTitle{KDRcpp}
}
\description{
\packageDescription{KDRcpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{KDRcpp}
\packageIndices{KDRcpp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{KDRcpp}

Maintainer: \packageMaintainer{KDRcpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
