% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RBFdot}
\alias{RBFdot}
\title{Radial basis function kernel matrix}
\usage{
RBFdot(X, Y, sigma)
}
\arguments{
\item{X}{an n x d matrix or numeric vector}

\item{Y}{an n x e matrix or numeric vector}

\item{sigma}{the scale parameter for the rbf kernel}
}
\value{
an n x n rbf kernel matrix
}
\description{
Radial basis function kernel matrix
}
\details{
this should be slightly faster than using \code{\link[kernlab]{kernelMatrix}}
}
\examples{
x <- as.matrix(rnorm(5e2, 0, 1))
K <- RBFdot(x, x, .5)
}

