% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distSquared}
\alias{distSquared}
\title{Squared distance matrix}
\usage{
distSquared(A, B)
}
\arguments{
\item{A}{an n x d matrix}

\item{B}{an n x d matrix}
}
\value{
an n x n distance matrix
}
\description{
Squared distance matrix computation
}

